begins host can view their reservations spec
