+++ spec/features/hosts/host_can_change_property_status_spec.rb
@@ -1,0 +1,40 @@
+require 'rails_helper'
+
+RSpec.feature "Host can change property status", type: :feature do
+  scenario "change status from unlisted to listed" do
+    role = create(:role, name: 'host')
+    host = create(:user)
+    host.roles << role
+
+    listing = create(:listing)
+    listing.user = host
+
+    allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(host)
+
+    visit host_dashboard_path(host)
+    click_on "View All Properties"
+
+    expect(current_path).to eq(host_properties_path)
+
+    within(".pending_properties") do
+      expect(page).to have_content(listing.name)
+      click_on "Activate"
+    end
+
+    expect(listing.status).to eq(1)
+  end
+
+  xit "traveler cannot change listing status" do
+    user = create(:user)
+    role = create(:role, name: 'traveler')
+
+
+  end
+
+  xit "traveler cannot see an unlisted listing" do
+    user = create(:user)
+    role = create(:role, name: 'traveler')
+
+
+  end
+end
