+++ spec/features/listings/guest_visits_homepage_spec.rb
