+++ spec/features/login-signup-logout/admin_can_log_in_spec.rb
