+++ spec/features/login-signup-logout/guest_creates_user_account_spec.rb
