+++ spec/features/login-signup-logout/guest_user_can_create_account_spec.rb
