+++ spec/features/login-signup-logout/user_can_login_spec.rb
