+++ spec/features/login-signup-logout/user_can_logout_spec.rb
