+++ spec/features/profile/user_can_view_own_profile_spec.rb
