+++ spec/features/search/guest_can_search_properties_spec.rb
